Fix #9281: fix duplication of get_datamodel_column_fqn
